Wakeshift is an app that uses data from the wakeshift mask
The Wakeshift Mask and Wakeshift App work conjunctively in order to reduce jet lag-related fatigue for space and Earth travelers alike, while also empowering users to maintain a balance of nutrition, fitness, and sleep. The eye mask contains an electroencephalogram (EEG) and reflective pulse oximetry sensors in the forehead area, as well as an electrooculography (EOG) sensor in the eye area. Together, the EEG and EOG detect the stage of sleep the user resides in and determines when and which binaural beats to play. The app consolidates nutrition, exercise, and sleep information to improve restfulness for sleep-deprived individuals.